refactor(rust): split commands/app, team_hub/mod, pty/session and unify path/error
