+++ src-tauri/src/util/mod.rs
@@ -1,2 +1,4 @@
 // Issue #140: ログ redaction とユーティリティの集約モジュール。
 pub mod log_redact;
+
+pub mod config_paths;
